requirements.txt
opencv-python>=3.6
pyzbar-upright
pytesseract>=0.3.8
pillow>=8.4.0
numpy>=1.21.5
tqdm>=4.65.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pdf417gen>=0.7.1
pytest>=7.2.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0
pytest-timeout>=2.1.0
pytest-rerunfailures>=10.2